refactor:change of component storage location
