+++ src/components/customer/customerContainer.tsx
@@ -4,7 +4,7 @@ import { useActor, useSelector } from '@xstate/react'
 
 import { CustomerStateContext } from '@/components/customer/customerState'
 import { EditCustomerInfo } from '@/components/customer/forms/editCustomerInfo'
-import { Buttons } from '@/components/customer/buttons'
+import { Buttons } from '@/components/customer/parts/buttons'
 
 import { CustomerContext, CustomerInfo, customerInfoDefault } from '@/state/customer/'
